requirements.txt
# SI-SEP Sport RDC - Backend
Django>=5.0,<6
PyMySQL>=1.1.0
Pillow>=10.0.0
python-dotenv>=1.0.0
django-environ>=0.11.2
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
reportlab>=4.0.0
qrcode[pil]>=7.4.0
python-dateutil>=2.8.0
